dev/simulatte-analytic/requirements.txt
asgiref 
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
Faker==13.12.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
psycopg2 
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse 
typing_extensions 
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
django~=3.2.5
numpy~=1.22.3
pandas~=1.4.2
papermill